+++ ship/LambdaShuttle.cpp
@@ -1,9 +1,11 @@
 #include "LambdaShuttle.hpp"
 
 const std::string LambdaShuttle::MODEL = "Lambda-class shuttle";
-const unsigned long long LambdaShuttle::WEIGHT = 360;
+const double LambdaShuttle::WEIGHT = 360;
 const unsigned LambdaShuttle::SPEED = 54;
 const double LambdaShuttle::MAX_LOAD = 80;
 unsigned LambdaShuttle::id = 0;
 
-LambdaShuttle::LambdaShuttle(double currentLoad) : LoadedShip(WEIGHT, SPEED, MODEL, ++id, MAX_LOAD, currentLoad) {}
+LambdaShuttle::LambdaShuttle(double currentLoad) : LoadedShip(WEIGHT, SPEED, MODEL,
+                                                              ++id, MAX_LOAD,
+                                                              currentLoad) {}
