+++ ship/LambdaShuttle.hpp
@@ -4,16 +4,16 @@
 #include "Ship.hpp"
 #include "LoadedShip.hpp"
 
-class LambdaShuttle: public LoadedShip {
+class LambdaShuttle : public LoadedShip {
 public:
     explicit LambdaShuttle(double currentLoad);
 
 private:
-   static const std::string MODEL;
-   static const unsigned long long WEIGHT;
-   static const unsigned SPEED;
-   static const double MAX_LOAD;
-   static unsigned id;
+    static const std::string MODEL;
+    static const double WEIGHT;
+    static const unsigned SPEED;
+    static const double MAX_LOAD;
+    static unsigned id;
 };
