+++ ship/StarDestroyer.hpp
@@ -6,14 +6,14 @@
 
 class StarDestroyer : public LoadedShip {
 public:
-   explicit StarDestroyer(double currentLoad);
+    explicit StarDestroyer(double currentLoad);
 
 private:
-   static const std::string MODEL;
-   static const unsigned long long WEIGHT;
-   static const unsigned SPEED;
-   static const double MAX_LOAD;
-   static unsigned id;
+    static const std::string MODEL;
+    static const double WEIGHT;
+    static const unsigned SPEED;
+    static const double MAX_LOAD;
+    static unsigned id;
 };
