+++ ship/TIELN.hpp
@@ -5,13 +5,13 @@
 
 class TIELN : public Ship {
 public:
-   TIELN();
+    TIELN();
 
 private:
-   static const std::string MODEL;
-   static const unsigned long long WEIGHT;
-   static const unsigned SPEED;
-   static unsigned id;
+    static const std::string MODEL;
+    static const double WEIGHT;
+    static const unsigned SPEED;
+    static unsigned id;
 };
